topic modeling/requirements.txt
aspose-words==24.11.0
nltk==3.9.1
gensim==4.3.3
pyLDAvis==3.4.1